Add action to resubmit image from profile Improve the on input change function Move functions away from App to containers Connect all containers to Redux Reduce number of props given to containers
