+++ src/constants.js
@@ -5,9 +5,10 @@ export const DATABASE = productionEnv ? 'https://fierce-oasis-21316.herokuapp.co
 
 // Image url setup
 export const CHANGE_INPUT_VALUE = 'CHANGE_INPUT_VALUE';
-export const RESUBMIT_INPUT_VALUE = 'RESUBMIT_INPUT_VALUE';
 export const SET_IMAGE_URL = 'SET_IMAGE_URL';
 export const SET_IMAGE_URL_ERROR = 'SET_IMAGE_URL_ERROR';
+export const RESUBMIT_INPUT_VALUE = 'RESUBMIT_INPUT_VALUE';
+export const RESUBMIT_IMAGE_URL = 'RESUBMIT_IMAGE_URL';
 
 // User information
 export const CHANGE_USER = 'CHANGE_USER';
